requirements.txt
openai>=1.3.0
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.10.6
sentence-transformers==3.3.1
faiss-cpu==1.10.0
numpy<3.0.0
metaphone==0.6
rapidfuzz==3.14.3
python-dotenv==1.0.1
openai==1.60.1
google-genai==0.2.1
scikit-learn==1.6.1
pymongo==4.11.1

pandas 
mlxtend
gensim
